+++ git-webui/release/share/git-webui/webui/js/git-webui.js
@@ -176,7 +176,6 @@ webui.git = function(cmd, arg1, arg2, arg3, arg4) {
                             location.reload();
                             return false;
                         }
-                        // here
                         webui.showError(displayMessage);
                     }
                 } else {
@@ -2379,7 +2378,7 @@ webui.ChangedFilesView = function(workspaceView, type, label) {
         var combinedFiles = files.concat(rmFiles);
 
         if(combinedFiles.length != 0){
-            webui.git("stash push -- " + combinedFiles, function(output){
+            webui.git("stash push --include-untracked -- " + combinedFiles, function(output){
                 webui.showSuccess(output);
                 workspaceView.update("stash");
             });